bank: rewrite status cache
The status cache has two main issues,

 (1) It's not particularly concurrent, and causes slowdowns when running
     with highly parallel bank execution.

 (2) It has unbounded memory consumption, and can cause out of memory
     conditions when it needs to store large numbers of transactions.

It is rewritten in Firedancer for performance.  The general design is
there are two operations which need to be highly concurrent, insert and
query, and everything else is rare and happens between slots and can
largely just lock the whole thing.

The base case for insert is optimized to two always-uncontended, and one
very lightly contended compare and swap.  Query is fully lockless.

The transaction result storage is combined between the snapshot service
cache, and the query lookup cache, which more than halves the memory
usage and the memory use is fixed up front.